bump version

+++ core/src/wire/scalars.rs
@@ -76,7 +76,7 @@ impl IntoWire for u64 {
     #[inline(always)]
     fn into_wire(self) -> WireType {
         let mut data = [0u8; 10];
-        let size = self.encode_var(&mut data) as u8;
+        let size = self.encode_var(&mut data);
         WireType::VarInt(data, size)
     }
 
@@ -105,7 +105,7 @@ impl IntoWire for i64 {
     #[inline(always)]
     fn into_wire(self) -> WireType {
         let mut data = [0u8; 10];
-        let size = self.encode_var(&mut data) as u8;
+        let size = self.encode_var(&mut data);
         WireType::VarInt(data, size)
     }
 
@@ -134,7 +134,7 @@ impl IntoWire for u32 {
     #[inline(always)]
     fn into_wire(self) -> WireType {
         let mut data = [0u8; 10];
-        let size = self.encode_var(&mut data) as u8;
+        let size = self.encode_var(&mut data);
         WireType::VarInt(data, size)
     }
 
@@ -163,7 +163,7 @@ impl IntoWire for i32 {
     #[inline(always)]
     fn into_wire(self) -> WireType {
         let mut data = [0u8; 10];
-        let size = self.encode_var(&mut data) as u8;
+        let size = self.encode_var(&mut data);
         WireType::VarInt(data, size)
     }
 
@@ -217,7 +217,7 @@ impl IntoWire for bool {
     #[inline(always)]
     fn into_wire(self) -> WireType {
         let mut data = [0u8; 10];
-        let size = if self { 1u32 } else { 0u32 }.encode_var(&mut data) as u8;
+        let size = if self { 1u32 } else { 0u32 }.encode_var(&mut data);
 
         WireType::VarInt(data, size)
     }
